prod_env/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
cachetools==5.5.2
certifi==2025.7.14
charset-normalizer==3.4.2
colorama==0.4.6
comm==0.2.3
debugpy==1.8.15
decorator==5.2.1
executing==2.2.0
filelock==3.18.0
fsspec==2025.7.0
google-auth==2.40.3
google-genai==1.27.0
gradio_client==1.11.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
ipykernel==6.30.0
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
joblib==1.5.1
jupyter_client==8.6.3
jupyter_core==5.8.1
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.3.2
packaging==25.0
pandas==2.3.1
parso==0.8.4
platformdirs==4.3.8
prompt_toolkit==3.0.51
psutil==7.0.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==311
PyYAML==6.0.2
pyzmq==27.0.0
requests==2.32.4
rsa==4.9.1
scikit-learn==1.7.1
scipy==1.16.1
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
tenacity==8.5.0
threadpoolctl==3.6.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
websockets==15.0.1
